gb10: cycle 52 step 11d — fix NVFP4 split BC16 tmp_out type + un-mute nvcc
Codex deep numerical review of NVFP4 attention kernels revealed a
2-cycle-old correctness bug:

`flash_attention_2_decode_nvfp4kv_split_bc16_kernel` (head_dim>256
path used by Gemma 4 31B global-attention layers at long context)
declared tmp_out as `__half*` while the helper
`write_split_empty_partition` and the reduce kernel
`paged_attention_reduce_f16_kernel` both expect `float*`. Cycle 21
widened the main split kernel + helper to f32 to avoid f16 round-trip
precision loss, but missed this BC16 variant.

Type mismatch is a hard nvcc compile error
("argument of type \"__half *\" is incompatible with parameter of
type \"float *\""), which was silently dropped because
kernels/build.sh suppressed nvcc stderr via `2>/dev/null`. The PTX
in kernels/sm_121/manifest.json was therefore STALE — last
successfully built before cycle 21 — and the BC16 split kernel
was either:
  (a) absent from the PTX (manifest sha matches an old build) →
      `kernels.fn_decode_nvfp4kv_split_bc16` returns None →
      head_dim>256 split decode falls through to bc=32 (which has
      the wrong smem budget for hd=512, undefined behavior), or
  (b) loaded with the OLD (pre-cycle-21) bc16 ABI writing __half
      tmp_out → reduce kernel reads f32 from the same buffer →
      garbage attention output.

Either path produces gibberish on Gemma 4 31B global layers (every
6th) at context length >= partition_size. Likely the actual
mechanism behind the historical "la la la" / weather-Korean
symptoms.

Fix:
  1. BC16 kernel signature: tmp_out: __half* → float*. Body already
     does scalar f32 stores via `tmp_out[...] = acc[r] * inv_sum`,
     so no further changes needed inside.
  2. PTX rebuilt: bytes 246308 → 243091, new sha 059829bf...
     (was 8a237582... — stale by ~2 hours of source mods).
  3. build.sh: removed `2>/dev/null` from the nvcc invocation. The
     next silent build error becomes loud; this class of bug
     surfaces immediately instead of after weeks of debugging.

CYCLE 52 STATUS:
  11a (1b9eab2): launch-shape validation
  11b+11c+11b.1 (bc358cf+3ae5e10): null guards + head_dim cross-check
  11d (this commit): NVFP4 split-BC16 type fix — likely fixes gibberish

Server --features gb10 builds clean. PTX manifest updated.
Smoke test against ebircak/gemma-4-31B-it-... or the existing
production NVFP4 KV path should now work correctly at long context;
v3/tools/kv_policy_matrix.sh remains the regression harness.

+++ kernels/flash_attention_split_decode_nvfp4kv.cu
@@ -393,7 +393,11 @@ __global__ void flash_attention_2_decode_nvfp4kv_split_kernel(
 
 extern "C"
 __global__ void flash_attention_2_decode_nvfp4kv_split_bc16_kernel(
-    __half*              __restrict__ tmp_out,
+    // tmp_out is f32 [S,H,P,D] — matches the BC=32 split kernel and
+    // the f32 contract `paged_attention_reduce_f16_kernel` expects.
+    // Cycle 21 widened the main split kernel to f32 but missed this
+    // BC16 variant; cycle 52 step 11d fix.
+    float*               __restrict__ tmp_out,
     float*               __restrict__ max_logits,
     float*               __restrict__ exp_sums,
     const unsigned char* __restrict__ query,
